refactor(user-payment-list): rename component
