+++ src/components/userPaymentList/ExpandableRowsComponent/ExpandableRowsComponent.helper.ts
