+++ src/components/userPaymentList/UserPaymentListView/UserPaymentListView.test.tsx
@@ -1,0 +1,9 @@
+import { render } from "setupTests";
+
+import PaymentListView from "./UserPaymentListView";
+
+describe("PaymentListView", () => {
+  it("renders with default props", () => {
+    render(<PaymentListView />);
+  });
+});
