+++ src/containers/UserPaymentList/UserPaymentList.types.ts
@@ -1,4 +1,4 @@
 // Interfaces and types from component AdminPaymentList
 
 // Container Props
-export interface AdminPaymentListContainerProps {}
+export interface UserPaymentListContainerProps {}
