--- src/pages/AdminPaymentListPage.tsx
@@ -1,3 +1,0 @@
-import AdminPaymentListContainer from "containers/AdminPaymentList/AdminPaymentList.container";
-
-export default AdminPaymentListContainer;
