+++ src/pages/UserPaymentListPage.tsx
@@ -1,0 +1,3 @@
+import UserPaymentListContainer from "containers/UserPaymentList/UserPaymentList.container";
+
+export default UserPaymentListContainer;
